fix(ui): make Next export template buildable across schemas
- Make template schema-agnostic (avoid literal narrowing from generated ths const)
- Wrap useSearchParams usage in Suspense via client/server split for static export
- Add smoke test that installs and builds the generated UI
